day154/day154-disaster-recovery/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
redis==5.0.1
aiohttp==3.9.1
asyncio==3.4.3
pytest==7.4.4
pytest-asyncio==0.23.3
structlog==24.1.0
prometheus-client==0.19.0
psutil==5.9.6
pyyaml==6.0.1
requests==2.31.0
websockets==12.0
python-multipart==0.0.6
pandas==2.1.4
plotly==5.18.0
numpy==1.26.3
aiofiles==23.2.1
